Crate breakpad_symbols

Structs

SimpleFrame

SimpleModule
A Module implementation that holds arbitrary data.

SimpleSymbolSupplier
Locate Breakpad text-format symbols in local disk paths.

SymbolFile
A parsed .sym file containing debug symbols.

Symbolizer
Symbolicate stack frames.

Enums

SymbolResult
Possible results of locating symbols.

Traits

FrameSymbolizer

Module
An executable or shared library loaded in a process.

SymbolSupplier
Locate symbols for a given module.

Functions

relative_symbol_path
Get a relative symbol path at which to locate symbols for module.