Function breakpad_symbols::relative_symbol_path

pub fn relative_symbol_path(module: &Module, extension: &str) -> Option<String>

Get a relative symbol path at which to locate symbols for module.

Symbols are generally stored in the layout used by Microsoft's symbol server and associated tools: <debug filename>/<debug identifier>/<debug filename>.sym. If debug filename ends with .pdb the leaf filename will have that removed. extension is the expected extension for the symbol filename, generally sym if Breakpad text format symbols are expected.

The debug filename and debug identifier can be found in the first line of the symbol file output by the dump_syms tool. You can use this script to run dump_syms and put the resulting symbol files in the proper directory structure.