Struct breakpad_symbols::SimpleFrame

pub struct SimpleFrame {
    pub instruction: u64,
    pub function: Option<String>,
    pub function_base: Option<u64>,
    pub source_file: Option<String>,
    pub source_line: Option<u32>,
    pub source_line_base: Option<u64>,
}

A simple implementation of FrameSymbolizer that just holds data.

Fields

instruction

The program counter value for this frame.

function

The name of the function in which the current instruction is executing.

function_base

The offset of the start of function from the module base.

source_file

The name of the source file in which the current instruction is executing.

source_line

The 1-based index of the line number in source_file in which the current instruction is executing.

source_line_base

The offset of the start of source_line from the function base.

Methods

impl SimpleFrame

fn with_instruction(instruction: u64) -> SimpleFrame

Instantiate a SimpleFrame with instruction pointer instruction.