Struct breakpad_symbols::SymbolFile

pub struct SymbolFile {
    pub files: HashMap<u32, String>,
    pub publics: Vec<PublicSymbol>,
    pub functions: RangeMap<Function>,
    pub cfi_stack_info: RangeMap<StackInfoCFI>,
    pub win_stack_info: RangeMap<StackInfoWin>,
}

A parsed .sym file containing debug symbols.

Fields

files

The set of source files involved in compilation.

publics

Publicly visible symbols.

functions

Functions.

cfi_stack_info

DWARF CFI unwind information.

win_stack_info

Windows unwind information.

Methods

impl SymbolFile

fn from_file(path: &Path) -> Result<SymbolFile, &'static str>

Parse a SymbolFile from path.

fn from_bytes(bytes: &[u8]) -> Result<SymbolFile, &'static str>

Parse an in-memory SymbolFile from bytes.

fn fill_symbol(&self, module: &Module, frame: &mut FrameSymbolizer)

Fill in as much source information for frame as possible.

fn find_nearest_public(&self, addr: u64) -> Option<&PublicSymbol>

Find the nearest PublicSymbol whose address is less than or equal to addr.