Struct breakpad_symbols::Symbolizer

pub struct Symbolizer {
    // some fields omitted
}

Symbolicate stack frames.

Methods

impl Symbolizer

fn new<T: SymbolSupplier + 'static>(supplier: T) -> Symbolizer

Create a Symbolizer that uses supplier to locate symbols.

fn get_symbol_at_address(&self, debug_file: &str, debug_id: &str, address: u64) -> Option<String>

Helper method for non-minidump-using callers.

Pass debug_file and debug_id describing a specific module, and address, a module-relative address, and get back a symbol in that module that covers that address, or None.

fn fill_symbol(&self, module: &Module, frame: &mut FrameSymbolizer)