Crate breakpad_symbols

A library for working with Google Breakpad's text-format symbol files.

The highest-level API provided by this crate is to use the Symbolizer struct.

Examples

use breakpad_symbols::{SimpleSymbolSupplier,Symbolizer,SimpleFrame,SimpleModule};
use std::path::PathBuf;
let paths = vec!(PathBuf::from("../testdata/symbols/"));
let supplier = SimpleSymbolSupplier::new(paths);
let symbolizer = Symbolizer::new(supplier);

// Simple function name lookup with debug file, debug id, address.
assert_eq!(symbolizer.get_symbol_at_address("test_app.pdb",
                                            "5A9832E5287241C1838ED98914E9B7FF1",
                                            0x1010)
              .unwrap(),
              "vswprintf");

Structs

SimpleFrame
A simple implementation of FrameSymbolizer that just holds data.

SimpleModule
A [Module][module] implementation that holds arbitrary data.

SimpleSymbolSupplier
An implementation of SymbolSupplier that loads Breakpad text-format symbols from local disk paths.

SymbolFile
A parsed .sym file containing debug symbols.

Symbolizer
Symbolicate stack frames.

Enums

SymbolResult
Possible results of locating symbols.

Traits

FrameSymbolizer
A trait for setting symbol information on something like a stack frame.

Module
An executable or shared library loaded in a process.

SymbolSupplier
A trait for things that can locate symbols for a given module.

Functions

relative_symbol_path
Get a relative symbol path at which to locate symbols for module.