Struct breakpad_symbols::SimpleModule

pub struct SimpleModule {
    pub base_address: Option<u64>,
    pub size: Option<u64>,
    pub code_file: Option<String>,
    pub code_identifier: Option<String>,
    pub debug_file: Option<String>,
    pub debug_id: Option<String>,
    pub version: Option<String>,
}

A Module implementation that holds arbitrary data.

This can be useful for getting symbols for a module when you have a debug id and filename but not an actual minidump. If you have a minidump, you should be using MinidumpModule.

Fields
base_address: Option<u64> size: Option<u64> code_file: Option<String> code_identifier: Option<String> debug_file: Option<String> debug_id: Option<String> version: Option<String>
Methods

impl SimpleModule

fn new(debug_file: &str, debug_id: &str) -> SimpleModule

Create a SimpleModule with the given debug_file and debug_id.

Uses default for the remaining fields.