Struct breakpad_symbols::SimpleModule

pub struct SimpleModule<'a> {
    pub base_address: Option<u64>,
    pub size: Option<u64>,
    pub code_file: Option<Cow<'a, str>>,
    pub code_identifier: Option<Cow<'a, str>>,
    pub debug_file: Option<Cow<'a, str>>,
    pub debug_id: Option<Cow<'a, str>>,
    pub version: Option<Cow<'a, str>>,
}

A Module implementation that holds arbitrary data.

This can be useful for getting symbols for a module when you have a debug id and filename but not an actual minidump.

Fields

base_address

size

code_file

code_identifier

debug_file

debug_id

version

Methods

impl<'a> SimpleModule<'a>

fn new(debug_file: &'a str, debug_id: &'a str) -> SimpleModule<'a>

Create a SimpleModule with the given debug_file and debug_id.

Uses default for the remaining fields.