Struct breakpad_symbols::SymbolFile

pub struct SymbolFile {
    pub files: HashMap<u32, String>,
    pub publics: Vec<PublicSymbol>,
    pub functions: RangeMap<u64, Function>,
    pub cfi_stack_info: RangeMap<u64, StackInfoCFI>,
    pub win_stack_framedata_info: RangeMap<u64, StackInfoWin>,
    pub win_stack_fpo_info: RangeMap<u64, StackInfoWin>,
}

A parsed .sym file containing debug symbols.

Fields
files: HashMap<u32, String>

The set of source files involved in compilation.

publics: Vec<PublicSymbol>

Publicly visible symbols.

functions: RangeMap<u64, Function>

Functions.

cfi_stack_info: RangeMap<u64, StackInfoCFI>

DWARF CFI unwind information.

win_stack_framedata_info: RangeMap<u64, StackInfoWin>

Windows unwind information (frame data).

win_stack_fpo_info: RangeMap<u64, StackInfoWin>

Windows unwind information (FPO data).

Methods

impl SymbolFile

pub fn from_file(path: &Path) -> Result<SymbolFile, &'static str>

Parse a SymbolFile from path.

pub fn from_bytes(bytes: &[u8]) -> Result<SymbolFile, &'static str>

Parse an in-memory SymbolFile from bytes.

pub fn fill_symbol(&self, module: &Module, frame: &mut FrameSymbolizer)

Fill in as much source information for frame as possible.

pub fn find_nearest_public(&self, addr: u64) -> Option<&PublicSymbol>

Find the nearest PublicSymbol whose address is less than or equal to addr.